refactor(vestad): consolidate timestamp and local-time helpers in time_utils, breaking the backup<->restic cycle
